requirements.txt
wxpython==4.2.0
pillow==9.4.0
google-cloud-translate==3.10.1
google-cloud-texttospeech==2.14.1
google-cloud-vision==3.3.1

future==0.18.3
py2exe==0.13.0.0
PyYAML==6.0
requests==2.28.2
keyboard==0.13.5
pyshadow==0.0.5
customtkinter==5.1.2

revChatGPT==3.1.4
baidu-aip==4.16.10
deepl-cli==0.3.1

mss==7.0.1
